add form validation HTML

+++ src/pages/jokes.tsx
@@ -1,16 +1,13 @@
 "use client"
 import React, { useEffect } from 'react'
 import { RouteComponentProps, useHistory } from 'react-router-dom';
-import { JokesPOST } from '../services/jokes';
-import useJokes from '../hooks/jokes';
 import { IJoke } from '../interfaces/jokes';
-import { useErrorBoundary } from 'react-error-boundary'
+import useJokes from '../hooks/jokes';
 
 export const Jokes = ({ match }: RouteComponentProps<{ jokeId?: string }>) => {
   const history = useHistory();
   const { jokeId } = match.params;
-  const { error, joke, getJoke, updateJoke, deleteJoke, createJoke } = useJokes();
-  const { showBoundary } = useErrorBoundary()
+  const { joke, getJoke, updateJoke, deleteJoke, createJoke } = useJokes();
 
   useEffect(() => {
     if (jokeId) {
@@ -30,9 +27,18 @@ export const Jokes = ({ match }: RouteComponentProps<{ jokeId?: string }>) => {
       views: parseInt(getValue(event, 'views')),
       body: getValue(event, 'body'),
       author: getValue(event, 'author'),
-      id: jokeId
+      id: jokeId,
     }
-    console.log('handleSubmit', { jokeId })
+
+    // const validJoke = JSON.parse(JSON.stringify(jokeForm))
+    // delete validJoke.id
+    // delete validJoke.createdAt
+    // const validated = Object.values(validJoke).every(value => !!value);
+    // if (!validated) {
+    //   alert('Please fill all fields')
+    //   return
+    // }
+
     if (jokeId) {
       jokeForm.createdAt = joke.createdAt
       await updateJoke(jokeForm)
@@ -47,31 +53,33 @@ export const Jokes = ({ match }: RouteComponentProps<{ jokeId?: string }>) => {
 
   const handleDelete = async (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
     event.preventDefault()
-    const res = await deleteJoke(jokeId as string)
-    if (res) history.push('/jokes')
+    if (window.confirm('Are you sure you want to delete this joke?')) {
+      const res = await deleteJoke(jokeId as string)
+      if (res) history.push('/jokes')
+    }
   }
 
   return (
     <form onSubmit={handleSubmit}>
       <div>
         <label htmlFor="title">Title:</label>
-        <input type="text" id="title" name="title" defaultValue={joke.title}></input>
+        <input type="text" id="title" name="title" required maxLength={50} defaultValue={joke.title}></input>
       </div>
       <div>
         <label htmlFor="body">Body:</label>
-        <input type="text" id="body" name="body" defaultValue={joke.body}></input>
+        <input type="text" id="body" name="body" required maxLength={350} defaultValue={joke.body}></input>
       </div>
       <div>
         <label htmlFor="author">Author:</label>
-        <input type="text" id="author" name="author" defaultValue={joke.author}></input>
+        <input type="text" id="author" name="author" required maxLength={30} defaultValue={joke.author}></input>
       </div>
-      <div>
+      {/* <div>
         <label htmlFor="date">Created date:</label>
         <input readOnly type="text" id="date" name="date" defaultValue={joke.createdAt}></input>
-      </div>
+      </div> */}
       <div>
         <label htmlFor="views">Views:</label>
-        <input type="text" id="views" name="views" defaultValue={joke.views}></input>
+        <input type="number" id="views" name="views" required defaultValue={joke.views}></input>
       </div>
       <div>
         <button type='submit'>Save</button>
